+++ conda/cli/activate.py
@@ -83,12 +83,15 @@ def _get_prefix_paths(prefix):
         yield os.path.join(prefix, 'bin')
 
 
-def binpath_from_arg(arg, shell):
+def binpath_from_arg(arg, shell, going_to_shell=True):
     shelldict = shells[shell] if shell else {}
     # prefix comes back as platform-native path
     prefix = prefix_from_arg(arg, shell)
     # convert paths to shell-native paths
-    return [shelldict['path_to'](path) for path in _get_prefix_paths(prefix)]
+    if going_to_shell:
+        return [shelldict['path_to'](path) for path in _get_prefix_paths(prefix)]
+    else:
+        return [path for path in _get_prefix_paths(prefix)]
 
 
 def pathlist_to_str(paths, escape_backslashes=True):
@@ -105,12 +108,15 @@ def pathlist_to_str(paths, escape_backslashes=True):
     return path
 
 
-def get_activate_path(prefix, shell):
+def get_activate_path(prefix, shell, going_to_shell=True):
     shelldict = shells[shell] if shell else {}
-    binpath = binpath_from_arg(prefix, shell)
+    binpath = binpath_from_arg(prefix, shell, going_to_shell)
 
     # prepend our new entries onto the existing path and make sure that the separator is native
-    path = shelldict['pathsep'].join(binpath)
+    if going_to_shell:
+        path = shelldict['pathsep'].join(binpath)
+    else:
+        path = os.pathsep.join(binpath)
     return path
 
 
@@ -146,11 +152,11 @@ def main():
                                              shell and env name".format(sys_argv[1]))
 
     if sys_argv[1] == '..activate':
-        print(get_activate_path(sys_argv[3], shell))
+        print(get_activate_path(sys_argv[3], shell, True))
         sys.exit(0)
 
     elif sys_argv[1] == '..deactivate.path':
-        activation_path = get_activate_path(sys_argv[3], shell)
+        activation_path = get_activate_path(sys_argv[3], shell, False)
 
         if os.getenv('_CONDA_HOLD'):
             new_path = regex.sub(r'%s(:?)' % regex.escape(activation_path),
@@ -160,6 +166,7 @@ def main():
             new_path = regex.sub(r'%s(:?)' % regex.escape(activation_path), r'',
                                  os.environ[str('PATH')], 1)
 
+        new_path = shells[shell]['path_to'](new_path)
         print(new_path)
         sys.exit(0)