+++ conda/cli/main_package.py
@@ -1,6 +1,3 @@
-# (c) Continuum Analytics, Inc. / http://continuum.io
-# All Rights Reserved
-#
 # conda is distributed under the terms of the BSD 3-clause license.
 # Consult LICENSE.txt or http://opensource.org/licenses/BSD-3-Clause.
 
@@ -17,56 +14,14 @@ import re
 import tarfile
 import tempfile
 
-from .conda_argparse import add_parser_prefix
+from .._vendor.auxlib.entity import EntityEncoder
 from ..base.context import context
 from ..common.compat import PY3, itervalues
-
-descr = "Low-level conda package utility. (EXPERIMENTAL)"
-
-
-def configure_parser(sub_parsers):
-    p = sub_parsers.add_parser(
-        'package',
-        description=descr,
-        help=descr,
-    )
-    add_parser_prefix(p)
-    p.add_argument(
-        '-w', "--which",
-        metavar="PATH",
-        nargs='+',
-        action="store",
-        help="Given some PATH print which conda package the file came from.",
-    )
-    p.add_argument(
-        '-r', "--reset",
-        action="store_true",
-        help="Remove all untracked files and exit.",
-    )
-    p.add_argument(
-        '-u', "--untracked",
-        action="store_true",
-        help="Display all untracked files and exit.",
-    )
-    p.add_argument(
-        "--pkg-name",
-        action="store",
-        default="unknown",
-        help="Package name of the created package.",
-    )
-    p.add_argument(
-        "--pkg-version",
-        action="store",
-        default="0.0",
-        help="Package version of the created package.",
-    )
-    p.add_argument(
-        "--pkg-build",
-        action="store",
-        default=0,
-        help="Package build number of the created package.",
-    )
-    p.set_defaults(func=execute)
+from ..common.path import paths_equal
+from ..core.linked_data import is_linked, linked, linked_data
+from ..gateways.disk.delete import rmtree
+from ..install import PREFIX_PLACEHOLDER
+from ..misc import untracked
 
 
 def remove(prefix, files):
@@ -87,7 +42,6 @@ def remove(prefix, files):
 
 
 def execute(args, parser):
-    from ..misc import untracked
 
     prefix = context.target_prefix
 
@@ -117,7 +71,6 @@ def execute(args, parser):
 
 
 def get_installed_version(prefix, name):
-    from ..core.linked_data import linked_data
     for info in itervalues(linked_data(prefix)):
         if info['name'] == name:
             return str(info['version'])
@@ -142,8 +95,6 @@ def create_info(name, version, build_number, requires_py):
 
 shebang_pat = re.compile(r'^#!.+$', re.M)
 def fix_shebang(tmp_dir, path):
-    from ..install import PREFIX_PLACEHOLDER
-
     if open(path, 'rb').read(2) != '#!':
         return False
 
@@ -170,7 +121,6 @@ def _add_info_dir(t, tmp_dir, files, has_prefix, info):
             fo.write(f + '\n')
 
     with open(join(info_dir, 'index.json'), 'w') as fo:
-        from .._vendor.auxlib.entity import EntityEncoder
         json.dump(info, fo, indent=2, sort_keys=True, cls=EntityEncoder)
 
     if has_prefix:
@@ -222,7 +172,6 @@ def create_conda_pkg(prefix, files, info, tar_path, update_info=None):
         update_info(info)
     _add_info_dir(t, tmp_dir, files, has_prefix, info)
     t.close()
-    from ..gateways.disk.delete import rmtree
     rmtree(tmp_dir)
     return warnings
 
@@ -230,7 +179,6 @@ def create_conda_pkg(prefix, files, info, tar_path, update_info=None):
 def make_tarbz2(prefix, name='unknown', version='0.0', build_number=0,
                 files=None):
     if files is None:
-        from ..misc import untracked
         files = untracked(prefix)
     print("# files: %d" % len(files))
     if len(files) == 0:
@@ -258,14 +206,12 @@ def which_package(path):
     the conda packages the file came from.  Usually the iteration yields
     only one package.
     """
-    from ..core.linked_data import is_linked, linked
     path = abspath(path)
     prefix = which_prefix(path)
     if prefix is None:
         from ..exceptions import CondaVerificationError
         raise CondaVerificationError("could not determine conda prefix from: %s" % path)
 
-    from ..common.path import paths_equal
     for dist in linked(prefix):
         meta = is_linked(prefix, dist)
         if any(paths_equal(join(prefix, f), path) for f in meta['files']):