+++ tests/test_resolve.py
@@ -1,9 +1,6 @@
 from __future__ import absolute_import, print_function
 
-from datetime import datetime
 import unittest
-
-import pytest
 
 from conda.base.context import context, reset_context
 from conda.common.compat import iteritems
@@ -13,10 +10,10 @@ from conda.models.channel import Channel
 from conda.models.dist import Dist
 from conda.models.index_record import IndexRecord
 from conda.resolve import MatchSpec, Resolve, ResolvePackageNotFound
-from tests.helpers import raises
-from .helpers import get_index_r_1
+from .helpers import get_index_r_1, get_index_r_3, raises
 
 index, r, = get_index_r_1()
+f_mkl = set(['mkl'])
 
 
 def add_defaults_if_no_channel(string):
@@ -157,11 +154,11 @@ def test_get_dists():
     assert Dist('channel-1::dynd-python-0.3.0-np17py33_0.tar.bz2') in dists
 
 
-def test_generate_eq():
+def test_generate_eq_1():
     reduced_index = r.get_reduced_index(['anaconda'])
     r2 = Resolve(reduced_index, True, True)
     C = r2.gen_clauses()
-    eqv, eqb = r2.generate_version_metrics(C, list(r2.groups.keys()))
+    eqc, eqv, eqb = r2.generate_version_metrics(C, list(r2.groups.keys()))
     # Should satisfy the following criteria:
     # - lower versions of the same package should should have higher
     #   coefficients.
@@ -170,8 +167,10 @@ def test_generate_eq():
     # - a package that only has one version should not appear, unless
     #   include=True as it will have a 0 coefficient. The same is true of the
     #   latest version of a package.
+    eqc = {Dist(key).to_filename(): value for key, value in iteritems(eqc)}
     eqv = {Dist(key).to_filename(): value for key, value in iteritems(eqv)}
     eqb = {Dist(key).to_filename(): value for key, value in iteritems(eqb)}
+    assert eqc == {}
     assert eqv == {
         'accelerate-1.0.0-np15py26_p0.tar.bz2': 2,
         'accelerate-1.0.0-np15py27_p0.tar.bz2': 2,
@@ -1109,7 +1108,7 @@ def test_remove():
         'zlib-1.2.7-0.tar.bz2']]
 
 
-def test_channel_priority():
+def test_channel_priority_1():
     channels = (
         Channel("channel-A"),
         Channel("channel-1"),
@@ -1157,6 +1156,292 @@ def test_channel_priority():
     assert installed1 != installed2
     assert installed1 != installed3
     assert installed2 == installed3
+
+
+def test_channel_priority_2():
+    this_index = index.copy()
+    index3, r3 = get_index_r_3()
+    this_index.update(index3)
+    spec = ['pandas', 'python 2.7*']
+    channels = (Channel('channel-1'), Channel('channel-3'))
+    this_r = Resolve(this_index, channels=channels)
+    with env_var("CONDA_CHANNEL_PRIORITY", "True", reset_context):
+        dists = this_r.get_reduced_index(spec)
+        r2 = Resolve(dists, True, True, channels=channels)
+        C = r2.gen_clauses()
+        eqc, eqv, eqb = r2.generate_version_metrics(C, list(r2.groups.keys()))
+        eqc = {str(Dist(key)): value for key, value in iteritems(eqc)}
+        assert eqc == {
+            'channel-3::openssl-1.0.1c-0': 1,
+            'channel-3::openssl-1.0.1g-0': 1,
+            'channel-3::openssl-1.0.1h-0': 1,
+            'channel-3::openssl-1.0.1h-1': 1,
+            'channel-3::openssl-1.0.1j-0': 1,
+            'channel-3::openssl-1.0.1j-1': 1,
+            'channel-3::openssl-1.0.1j-2': 1,
+            'channel-3::openssl-1.0.1j-3': 1,
+            'channel-3::openssl-1.0.1j-4': 1,
+            'channel-3::openssl-1.0.1j-5': 1,
+            'channel-3::openssl-1.0.1k-0': 1,
+            'channel-3::openssl-1.0.1k-1': 1,
+            'channel-3::openssl-1.0.2d-0': 1,
+            'channel-3::openssl-1.0.2e-0': 1,
+            'channel-3::openssl-1.0.2f-0': 1,
+            'channel-3::openssl-1.0.2g-0': 1,
+            'channel-3::openssl-1.0.2h-0': 1,
+            'channel-3::openssl-1.0.2h-1': 1,
+            'channel-3::openssl-1.0.2i-0': 1,
+            'channel-3::openssl-1.0.2j-0': 1,
+            'channel-3::openssl-1.0.2k-0': 1,
+            'channel-3::openssl-1.0.2k-1': 1,
+            'channel-3::openssl-1.0.2k-2': 1,
+            'channel-3::openssl-1.0.2l-0': 1,
+            'channel-3::python-2.7.10-0': 1,
+            'channel-3::python-2.7.10-1': 1,
+            'channel-3::python-2.7.10-2': 1,
+            'channel-3::python-2.7.11-0': 1,
+            'channel-3::python-2.7.11-5': 1,
+            'channel-3::python-2.7.12-0': 1,
+            'channel-3::python-2.7.12-1': 1,
+            'channel-3::python-2.7.13-0': 1,
+            'channel-3::python-2.7.3-2': 1,
+            'channel-3::python-2.7.3-3': 1,
+            'channel-3::python-2.7.3-4': 1,
+            'channel-3::python-2.7.3-5': 1,
+            'channel-3::python-2.7.3-6': 1,
+            'channel-3::python-2.7.3-7': 1,
+            'channel-3::python-2.7.4-0': 1,
+            'channel-3::python-2.7.5-0': 1,
+            'channel-3::python-2.7.5-1': 1,
+            'channel-3::python-2.7.5-2': 1,
+            'channel-3::python-2.7.5-3': 1,
+            'channel-3::python-2.7.6-0': 1,
+            'channel-3::python-2.7.6-1': 1,
+            'channel-3::python-2.7.6-2': 1,
+            'channel-3::python-2.7.7-0': 1,
+            'channel-3::python-2.7.7-2': 1,
+            'channel-3::python-2.7.8-0': 1,
+            'channel-3::python-2.7.8-1': 1,
+            'channel-3::python-2.7.9-0': 1,
+            'channel-3::python-2.7.9-1': 1,
+            'channel-3::python-2.7.9-2': 1,
+            'channel-3::python-2.7.9-3': 1,
+            'channel-3::readline-6.2-0': 1,
+            'channel-3::readline-6.2-2': 1,
+            'channel-3::sqlite-3.13.0-0': 1,
+            'channel-3::sqlite-3.7.13-0': 1,
+            'channel-3::sqlite-3.8.4.1-0': 1,
+            'channel-3::sqlite-3.8.4.1-1': 1,
+            'channel-3::sqlite-3.9.2-0': 1,
+            'channel-3::system-5.8-0': 1,
+            'channel-3::system-5.8-1': 1,
+            'channel-3::system-5.8-2': 1,
+            'channel-3::tk-8.5.13-0': 1,
+            'channel-3::tk-8.5.15-0': 1,
+            'channel-3::tk-8.5.18-0': 1,
+            'channel-3::zlib-1.2.7-0': 1,
+            'channel-3::zlib-1.2.7-1': 1,
+            'channel-3::zlib-1.2.7-2': 1,
+            'channel-3::zlib-1.2.8-0': 1,
+            'channel-3::zlib-1.2.8-3': 1,
+        }
+        installed_w_priority = [str(d) for d in this_r.install(spec)]
+        assert installed_w_priority == [
+            'channel-1::dateutil-2.1-py27_1',
+            'channel-1::numpy-1.7.1-py27_0',
+            'channel-1::openssl-1.0.1c-0',
+            'channel-1::pandas-0.11.0-np17py27_1',
+            'channel-1::python-2.7.5-0',
+            'channel-1::pytz-2013b-py27_0',
+            'channel-1::readline-6.2-0',
+            'channel-1::scipy-0.12.0-np17py27_0',
+            'channel-1::six-1.3.0-py27_0',
+            'channel-1::sqlite-3.7.13-0',
+            'channel-1::system-5.8-1',
+            'channel-1::tk-8.5.13-0',
+            'channel-1::zlib-1.2.7-0',
+        ]
+
+    with env_var("CONDA_CHANNEL_PRIORITY", "False", reset_context):
+        dists = this_r.get_reduced_index(spec)
+        r2 = Resolve(dists, True, True, channels=channels)
+        C = r2.gen_clauses()
+        eqc, eqv, eqb = r2.generate_version_metrics(C, list(r2.groups.keys()))
+        eqc = {str(Dist(key)): value for key, value in iteritems(eqc)}
+        assert eqc == {
+            'channel-1::accelerate-1.0.0-np16py27_p0': 2,
+            'channel-1::accelerate-1.0.0-np17py27_p0': 2,
+            'channel-1::accelerate-1.0.1-np16py27_p0': 1,
+            'channel-1::accelerate-1.0.1-np17py27_p0': 1,
+            'channel-1::bitarray-0.8.0-py27_0': 1,
+            'channel-1::dateutil-1.5-py27_0': 1,
+            'channel-1::llvmpy-0.11.1-py27_0': 1,
+            'channel-1::mkl-10.3-0': 1,
+            'channel-1::mkl-10.3-p1': 1,
+            'channel-1::mkl-10.3-p2': 1,
+            'channel-1::nose-1.1.2-py27_0': 2,
+            'channel-1::nose-1.2.1-py27_0': 1,
+            'channel-1::numba-0.7.1-np16py27_0': 1,
+            'channel-1::numba-0.7.1-np17py27_0': 1,
+            'channel-1::numbapro-0.10.0-np16py27_p0': 2,
+            'channel-1::numbapro-0.10.0-np17py27_p0': 2,
+            'channel-1::numbapro-0.10.1-np16py27_p0': 1,
+            'channel-1::numbapro-0.10.1-np17py27_p0': 1,
+            'channel-1::numexpr-2.0.1-np16py27_1': 1,
+            'channel-1::numexpr-2.0.1-np16py27_2': 1,
+            'channel-1::numexpr-2.0.1-np16py27_3': 1,
+            'channel-1::numexpr-2.0.1-np16py27_ce0': 1,
+            'channel-1::numexpr-2.0.1-np16py27_p1': 1,
+            'channel-1::numexpr-2.0.1-np16py27_p2': 1,
+            'channel-1::numexpr-2.0.1-np16py27_p3': 1,
+            'channel-1::numexpr-2.0.1-np16py27_pro0': 1,
+            'channel-1::numexpr-2.0.1-np17py27_1': 1,
+            'channel-1::numexpr-2.0.1-np17py27_2': 1,
+            'channel-1::numexpr-2.0.1-np17py27_3': 1,
+            'channel-1::numexpr-2.0.1-np17py27_ce0': 1,
+            'channel-1::numexpr-2.0.1-np17py27_p1': 1,
+            'channel-1::numexpr-2.0.1-np17py27_p2': 1,
+            'channel-1::numexpr-2.0.1-np17py27_p3': 1,
+            'channel-1::numexpr-2.0.1-np17py27_pro0': 1,
+            'channel-1::numpy-1.6.2-py27_1': 4,
+            'channel-1::numpy-1.6.2-py27_3': 4,
+            'channel-1::numpy-1.6.2-py27_4': 4,
+            'channel-1::numpy-1.6.2-py27_ce0': 4,
+            'channel-1::numpy-1.6.2-py27_p1': 4,
+            'channel-1::numpy-1.6.2-py27_p3': 4,
+            'channel-1::numpy-1.6.2-py27_p4': 4,
+            'channel-1::numpy-1.6.2-py27_pro0': 4,
+            'channel-1::numpy-1.7.0-py27_0': 1,
+            'channel-1::numpy-1.7.0-py27_p0': 1,
+            'channel-1::numpy-1.7.0b2-py27_ce0': 3,
+            'channel-1::numpy-1.7.0b2-py27_pro0': 3,
+            'channel-1::numpy-1.7.0rc1-py27_0': 2,
+            'channel-1::numpy-1.7.0rc1-py27_p0': 2,
+            'channel-1::openssl-1.0.1c-0': 13,
+            'channel-1::pandas-0.10.0-np16py27_0': 2,
+            'channel-1::pandas-0.10.0-np17py27_0': 2,
+            'channel-1::pandas-0.10.1-np16py27_0': 1,
+            'channel-1::pandas-0.10.1-np17py27_0': 1,
+            'channel-1::pandas-0.8.1-np16py27_0': 5,
+            'channel-1::pandas-0.8.1-np17py27_0': 5,
+            'channel-1::pandas-0.9.0-np16py27_0': 4,
+            'channel-1::pandas-0.9.0-np17py27_0': 4,
+            'channel-1::pandas-0.9.1-np16py27_0': 3,
+            'channel-1::pandas-0.9.1-np17py27_0': 3,
+            'channel-1::python-2.7.3-2': 10,
+            'channel-1::python-2.7.3-3': 10,
+            'channel-1::python-2.7.3-4': 10,
+            'channel-1::python-2.7.3-5': 10,
+            'channel-1::python-2.7.3-6': 10,
+            'channel-1::python-2.7.3-7': 10,
+            'channel-1::python-2.7.4-0': 9,
+            'channel-1::python-2.7.5-0': 8,
+            'channel-1::pytz-2012d-py27_0': 2,
+            'channel-1::pytz-2012j-py27_0': 1,
+            'channel-1::scikit-learn-0.13-np16py27_0': 1,
+            'channel-1::scikit-learn-0.13-np16py27_1': 1,
+            'channel-1::scikit-learn-0.13-np16py27_p0': 1,
+            'channel-1::scikit-learn-0.13-np16py27_p1': 1,
+            'channel-1::scikit-learn-0.13-np17py27_0': 1,
+            'channel-1::scikit-learn-0.13-np17py27_1': 1,
+            'channel-1::scikit-learn-0.13-np17py27_p0': 1,
+            'channel-1::scikit-learn-0.13-np17py27_p1': 1,
+            'channel-1::scipy-0.11.0-np16py27_2': 1,
+            'channel-1::scipy-0.11.0-np16py27_3': 1,
+            'channel-1::scipy-0.11.0-np16py27_ce1': 1,
+            'channel-1::scipy-0.11.0-np16py27_p2': 1,
+            'channel-1::scipy-0.11.0-np16py27_p3': 1,
+            'channel-1::scipy-0.11.0-np16py27_pro0': 1,
+            'channel-1::scipy-0.11.0-np16py27_pro1': 1,
+            'channel-1::scipy-0.11.0-np17py27_2': 1,
+            'channel-1::scipy-0.11.0-np17py27_3': 1,
+            'channel-1::scipy-0.11.0-np17py27_ce0': 1,
+            'channel-1::scipy-0.11.0-np17py27_ce1': 1,
+            'channel-1::scipy-0.11.0-np17py27_p2': 1,
+            'channel-1::scipy-0.11.0-np17py27_p3': 1,
+            'channel-1::scipy-0.11.0-np17py27_pro0': 1,
+            'channel-1::scipy-0.11.0-np17py27_pro1': 1,
+            'channel-1::six-1.2.0-py27_0': 1,
+            'channel-1::sqlite-3.7.13-0': 3,
+            'channel-1::tk-8.5.13-0': 2,
+            'channel-1::zlib-1.2.7-0': 1,
+            'channel-3::openssl-1.0.1c-0': 13,
+            'channel-3::openssl-1.0.1g-0': 12,
+            'channel-3::openssl-1.0.1h-0': 11,
+            'channel-3::openssl-1.0.1h-1': 11,
+            'channel-3::openssl-1.0.1j-0': 10,
+            'channel-3::openssl-1.0.1j-1': 10,
+            'channel-3::openssl-1.0.1j-2': 10,
+            'channel-3::openssl-1.0.1j-3': 10,
+            'channel-3::openssl-1.0.1j-4': 10,
+            'channel-3::openssl-1.0.1j-5': 10,
+            'channel-3::openssl-1.0.1k-0': 9,
+            'channel-3::openssl-1.0.1k-1': 9,
+            'channel-3::openssl-1.0.2d-0': 8,
+            'channel-3::openssl-1.0.2e-0': 7,
+            'channel-3::openssl-1.0.2f-0': 6,
+            'channel-3::openssl-1.0.2g-0': 5,
+            'channel-3::openssl-1.0.2h-0': 4,
+            'channel-3::openssl-1.0.2h-1': 4,
+            'channel-3::openssl-1.0.2i-0': 3,
+            'channel-3::openssl-1.0.2j-0': 2,
+            'channel-3::openssl-1.0.2k-0': 1,
+            'channel-3::openssl-1.0.2k-1': 1,
+            'channel-3::openssl-1.0.2k-2': 1,
+            'channel-3::python-2.7.10-0': 3,
+            'channel-3::python-2.7.10-1': 3,
+            'channel-3::python-2.7.10-2': 3,
+            'channel-3::python-2.7.11-0': 2,
+            'channel-3::python-2.7.11-5': 2,
+            'channel-3::python-2.7.12-0': 1,
+            'channel-3::python-2.7.12-1': 1,
+            'channel-3::python-2.7.3-2': 10,
+            'channel-3::python-2.7.3-3': 10,
+            'channel-3::python-2.7.3-4': 10,
+            'channel-3::python-2.7.3-5': 10,
+            'channel-3::python-2.7.3-6': 10,
+            'channel-3::python-2.7.3-7': 10,
+            'channel-3::python-2.7.4-0': 9,
+            'channel-3::python-2.7.5-0': 8,
+            'channel-3::python-2.7.5-1': 8,
+            'channel-3::python-2.7.5-2': 8,
+            'channel-3::python-2.7.5-3': 8,
+            'channel-3::python-2.7.6-0': 7,
+            'channel-3::python-2.7.6-1': 7,
+            'channel-3::python-2.7.6-2': 7,
+            'channel-3::python-2.7.7-0': 6,
+            'channel-3::python-2.7.7-2': 6,
+            'channel-3::python-2.7.8-0': 5,
+            'channel-3::python-2.7.8-1': 5,
+            'channel-3::python-2.7.9-0': 4,
+            'channel-3::python-2.7.9-1': 4,
+            'channel-3::python-2.7.9-2': 4,
+            'channel-3::python-2.7.9-3': 4,
+            'channel-3::sqlite-3.7.13-0': 3,
+            'channel-3::sqlite-3.8.4.1-0': 2,
+            'channel-3::sqlite-3.8.4.1-1': 2,
+            'channel-3::sqlite-3.9.2-0': 1,
+            'channel-3::tk-8.5.13-0': 2,
+            'channel-3::tk-8.5.15-0': 1,
+            'channel-3::zlib-1.2.7-0': 1,
+            'channel-3::zlib-1.2.7-1': 1,
+            'channel-3::zlib-1.2.7-2': 1,
+        }
+        installed_wo_priority = set([str(d) for d in this_r.install(spec)])
+        assert installed_wo_priority == {
+            'channel-3::openssl-1.0.2l-0',
+            'channel-3::python-2.7.13-0',
+            'channel-3::sqlite-3.13.0-0',
+            'channel-3::tk-8.5.18-0',
+            'channel-3::zlib-1.2.8-3',
+            'channel-1::dateutil-2.1-py27_1',
+            'channel-1::numpy-1.7.1-py27_0',
+            'channel-1::pandas-0.11.0-np17py27_1',
+            'channel-1::pytz-2013b-py27_0',
+            'channel-1::readline-6.2-0',
+            'channel-1::scipy-0.12.0-np17py27_0',
+            'channel-1::six-1.3.0-py27_0',
+        }
 
 
 def test_dependency_sort():