+++ CHANGELOG.md
@@ -119,6 +119,7 @@
 * remove support for with_features_depends (#5191)
 * resolve #5468 remove --alt-hint from CLI API (#5469)
 * resolve #5834 change default value of 'allow_softlinks' from True to False (#5835)
+* resolve #5842 add deprecation warnings for 'conda env upload' and 'conda env attach' (#5843)
 
 ### API
 * Add Solver from conda.core.solver with three methods to conda.api (4.4.0rc1) (#5838)
@@ -163,6 +164,9 @@
 * index building optimizations (#5776)
 * fix #5811 change safety_checks default to 'warn' for conda 4.4 (4.4.0rc1) (#5824)
 * add constrained dependencies to conda's own recipe (4.4.0rc1) (#5823)
+* clean up parser imports (4.4.0rc2) (#5844)
+* resolve #5983 add --download-only flag to create, install, and update (4.4.0rc2) (#5988)
+* add ca-certificates and certifi to aggressive_update_packages default (4.4.0rc2) (#5994)
 
 ### Bug Fixes
 * fix some conda-build compatibility issues (#5089)
@@ -212,6 +216,38 @@
 * rename PackageNotFoundError to PackagesNotFoundError and fix message formatting (#5602)
 * update conda 4.4 bld.bat windows recipe (#5573)
 * remove last remnant of CondaEnvRuntimeError (#5643)
+
+
+## 4.3.28 (unreleased)
+
+### Bug Fixes
+
+* fix #5854 remove imports of pkg_resources (#5991)
+* fix millisecond timestamps (#6001)
+
+
+## 4.3.27 (2017-09-18)
+
+### Bug Fixes
+
+* fix #5980 always delete_prefix_from_linked_data in rm_rf (#5982)
+
+
+## 4.3.26 (2017-09-15)
+
+### Deprecations/Breaking Changes
+* resolve #5922 prioritize channels within multi-channels (#5923)
+* add https://repo.continuum.io/pkgs/main to defaults multi-channel (#5931)
+
+### Improvements
+* add a channel priority minimization pass to solver logic (#5859)
+* invoke cmd.exe with /D for pre/post link/unlink scripts (#5926)
+* add boto3 use to s3 adapter (#5949)
+
+### Bug Fixes
+* always remove linked prefix entry with rm_rf (#5846)
+* resolve #5920 bump repodata pickle version (#5921)
+* fix msys2 activate and deactivate (#5950)
 
 
 ## 4.3.25 (2017-08-16)